+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_map_example/pages/moving_markers.dart';
 import './pages/esri.dart';
 import './pages/home.dart';
 import './pages/map_controller.dart';
@@ -8,6 +9,7 @@ import './pages/plugin_api.dart';
 import './pages/polyline.dart';
 import './pages/tap_to_add.dart';
 import './pages/offline_map.dart';
+import './pages/on_tap.dart';
 
 void main() => runApp(new MyApp());
 
@@ -30,6 +32,8 @@ class MyApp extends StatelessWidget {
         MarkerAnchorPage.route: (context) => new MarkerAnchorPage(),
         PluginPage.route: (context) => new PluginPage(),
         OfflineMapPage.route: (context) => new OfflineMapPage(),
+        OnTapPage.route: (context) => new OnTapPage(),
+        MovingMarkersPage.route: (context) => new MovingMarkersPage(),
       },
     );
   }