+++ lib/src/layer/layer.dart
@@ -1,4 +1,5 @@
 class LayerOptions {
-
+  Stream<Null> rebuild;
+  LayerOptions({this.rebuild});
 }