+++ lib/src/layer/tile_layer.dart
@@ -10,6 +10,7 @@ import 'package:flutter_map/src/map/map.dart';
 import 'package:latlong/latlong.dart';
 import 'package:transparent_image/transparent_image.dart';
 import 'package:tuple/tuple.dart';
+import 'package:flutter_image/network.dart';
 
 import 'layer.dart';
 
@@ -24,15 +25,22 @@ class TileLayerOptions extends LayerOptions {
   ///
   ///https://a.tile.openstreetmap.org/12/2177/1259.png
   final String urlTemplate;
+
+  /// If `true`, inverses Y axis numbering for tiles (turn this on for
+  /// [TMS](https://en.wikipedia.org/wiki/Tile_Map_Service) services).
+  final bool tms;
+
   ///Size for the tile.
   ///Default is 256
   final double tileSize;
+
   ///Determiantes the max zoom applicable.
   ///In most tile providers goes from 0 to 19.
   final double maxZoom;
 
   final bool zoomReverse;
   final double zoomOffset;
+
   ///List of subdomains for the URL.
   ///
   ///Example:
@@ -49,6 +57,7 @@ class TileLayerOptions extends LayerOptions {
   ///https://b.tile.openstreetmap.org/{z}/{x}/{y}.png
   ///https://c.tile.openstreetmap.org/{z}/{x}/{y}.png
   final List<String> subdomains;
+
   ///Color shown behind the tiles.
   final Color backgroundColor;
 
@@ -81,29 +90,33 @@ class TileLayerOptions extends LayerOptions {
   ImageProvider placeholderImage;
   Map<String, String> additionalOptions;
 
-  TileLayerOptions({
-    this.urlTemplate,
-    this.tileSize = 256.0,
-    this.maxZoom = 18.0,
-    this.zoomReverse = false,
-    this.zoomOffset = 0.0,
-    this.additionalOptions = const <String, String>{},
-    this.subdomains = const <String>[],
-    this.keepBuffer = 2,
-    this.backgroundColor = const Color(0xFFE0E0E0), // grey[300]
-    this.placeholderImage,
-    this.offlineMode = false,
-    this.fromAssets = true,
-  });
+  TileLayerOptions(
+      {this.urlTemplate,
+      this.tileSize = 256.0,
+      this.maxZoom = 18.0,
+      this.zoomReverse = false,
+      this.zoomOffset = 0.0,
+      this.additionalOptions = const <String, String>{},
+      this.subdomains = const <String>[],
+      this.keepBuffer = 2,
+      this.backgroundColor = const Color(0xFFE0E0E0), // grey[300]
+      this.placeholderImage,
+      this.offlineMode = false,
+      this.tms = false,
+      this.fromAssets = true,
+      rebuild})
+      : super(rebuild: rebuild);
 }
 
 class TileLayer extends StatefulWidget {
   final TileLayerOptions options;
   final MapState mapState;
+  final Stream<Null> stream;
 
   TileLayer({
     this.options,
     this.mapState,
+    this.stream,
   });
 
   State<StatefulWidget> createState() {
@@ -128,7 +141,7 @@ class _TileLayerState extends State<TileLayer> {
   void initState() {
     super.initState();
     _resetView();
-    _moveSub = map.onMoved.listen((_) => _handleMove());
+    _moveSub = widget.stream.listen((_) => _handleMove());
   }
 
   void dispose() {
@@ -150,9 +163,16 @@ class _TileLayerState extends State<TileLayer> {
       'z': coords.z.round().toString(),
       's': _getSubdomain(coords)
     };
+    if(this.options.tms) {
+      data['y'] = _invertY(coords.y.round(), coords.z.round()).toString();
+    }
     Map<String, String> allOpts = new Map.from(data)
       ..addAll(this.options.additionalOptions);
     return util.template(this.options.urlTemplate, allOpts);
+  }
+
+  int _invertY(int y, int z) {
+    return ((1 << z) - 1) - y;
   }
 
   void _resetView() {
@@ -446,7 +466,7 @@ class _TileLayerState extends State<TileLayer> {
         return new FileImage(new File(url));
       }
     } else {
-      return new NetworkImage(url);
+      return new NetworkImageWithRetry(url);
     }
   }