+++ lib/src/map/flutter_map_state.dart
@@ -2,10 +2,14 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/src/core/point.dart';
 import 'package:flutter_map/src/gestures/gestures.dart';
+import 'package:flutter_map/src/layer/group_layer.dart';
 import 'package:flutter_map/src/map/map.dart';
+import 'package:positioned_tap_detector/positioned_tap_detector.dart';
+import 'package:async/async.dart';
 
 class FlutterMapState extends MapGestureMixin {
   final MapControllerImpl mapController;
+  final List<StreamGroup<Null>> groups = <StreamGroup<Null>>[];
   MapOptions get options => widget.options ?? new MapOptions();
   MapState mapState;
 
@@ -17,7 +21,31 @@ class FlutterMapState extends MapGestureMixin {
     mapController.state = mapState;
   }
 
+  void _dispose() {
+    groups.forEach((group) => group.close());
+    groups.clear();
+  }
+
+
+  @override
+  void dispose() {
+    _dispose();
+    super.dispose();
+  }
+
+  Stream<Null> _merge(LayerOptions options) {
+    if(options?.rebuild == null)
+      return mapState.onMoved;
+
+    StreamGroup<Null> group = new StreamGroup<Null>();
+    group.add(mapState.onMoved);
+    group.add(options.rebuild);
+    groups.add(group);
+    return group.stream;
+  }
+
   Widget build(BuildContext context) {
+    _dispose();
     return new LayoutBuilder(
         builder: (BuildContext context, BoxConstraints constraints) {
       mapState.size =
@@ -25,15 +53,19 @@ class FlutterMapState extends MapGestureMixin {
       var layerWidgets = widget.layers
           .map((layer) => _createLayer(layer, widget.options.plugins))
           .toList();
-      return new GestureDetector(
-        onScaleStart: handleScaleStart,
-        onScaleUpdate: handleScaleUpdate,
-        onScaleEnd: handleScaleEnd,
-        onTapUp: handleTapUp,
+      return PositionedTapDetector(
+        onTap: handleTap,
         onDoubleTap: handleDoubleTap,
-        child: new Container(
-          child: new Stack(
-            children: layerWidgets,
+        child: new GestureDetector(
+          onScaleStart: handleScaleStart,
+          onScaleUpdate: handleScaleUpdate,
+          onScaleEnd: handleScaleEnd,
+          child: new Container(
+          width: constraints.maxWidth,
+          height: constraints.maxHeight,
+            child: new Stack(
+              children: layerWidgets,
+            ),
           ),
         ),
       );
@@ -42,17 +74,26 @@ class FlutterMapState extends MapGestureMixin {
 
   Widget _createLayer(LayerOptions options, List<MapPlugin> plugins) {
     if (options is TileLayerOptions) {
-      return new TileLayer(options: options, mapState: mapState);
+      return new TileLayer(options: options, mapState: mapState, stream: _merge(options));
     }
     if (options is MarkerLayerOptions) {
-      return new MarkerLayer(options, mapState);
+      return new MarkerLayer(options, mapState, _merge(options));
     }
     if (options is PolylineLayerOptions) {
-      return new PolylineLayer(options, mapState);
+      return new PolylineLayer(options, mapState, _merge(options));
+    }
+    if (options is PolygonLayerOptions) {
+      return new PolygonLayer(options, mapState, _merge(options));
+    }
+    if (options is CircleLayerOptions) {
+      return new CircleLayer(options, mapState, _merge(options));
+    }
+    if (options is GroupLayerOptions) {
+      return new GroupLayer(options, mapState, _merge(options));
     }
     for (var plugin in plugins) {
       if (plugin.supportsLayer(options)) {
-        return plugin.createLayer(options, mapState);
+        return plugin.createLayer(options, mapState, _merge(options));
       }
     }
     return null;