+++ CHANGELOG.md
@@ -1,3 +1,47 @@
+## [0.4.0] - 12/31/2018
+- Zoom to focal point on double tap and scale gestures (#121)
+- Make anchor field public (#172)
+- FitBoundsOptions now uses EdgeInsets padding
+- Add GroupLayer
+- Update README
+
+Thanks to @tomwyr, @csjames, @kengu, @ocularrhythm for this release!
+
+## [0.3.0] - 11/1/2018
+- PositionCallback now has hasGesture #139
+
+Thanks to @gimox for this release!
+
+## [0.2.0] - 10/25/2018
+- Use NetworkImageWithRetry for tile layers (#145)
+- Add rebuild capability to LayerOptions (#144)
+- Added Circle layer (#137)
+- Prevent Map Layer Excessive Rebuilds (#131)
+
+Thanks to @kengu, @mortenboye, and @tomwyr for this release!
+
+## [0.1.4] - 9/24/2018
+- Polygon Support (#118)
+
+Thanks to @JulianBerger for this release!
+
+## [0.1.3] - 9/18/2018
+- fix identical map position callbacks (#111)
+- Prune tiles bug fix (#112)
+
+Thanks to @IhorKlimov and @tomwyr for this release!
+
+## [0.1.2] - 8/21/2018
+- Added polyline customisation options (#94)
+- Expose map bounds (#99)
+- Added onTap example (#103)
+- route bugfix (#104)
+- options is now required (#105)
+- Project refactor and changes to offline map #85
+
+Thanks to @LJaraCastillo, @ubilabs, @xqwzts, @vinicentus, and @lsaudon for this
+release!
+
 ## [0.1.0] - 8/21/2018
 - Set Dart SDK to ">=2.0.0 <3.0.0"